PID Values and sensor Measurements
Changed PID values and sensor Measurements

+++ STM32Cube/Components/Inc/hcsr04_sensor.h
@@ -18,7 +18,7 @@ extern "C" {
 /* Defines and variables ------------------------------------------*/
 
 #define MAX_SENSORS 2  /**< Maximum number of sensors. */
-typedef uint32_t TIM_Channel;
+typedef float TIM_Channel;
 
 /**
   * @brief  The sensor data and configurations structures.
@@ -28,7 +28,7 @@ struct us_sensor_str
     TIM_HandleTypeDef *htim_echo;   /**< Timer handle for echo. */
     TIM_HandleTypeDef *htim_trig;   /**< Timer handle for trigger. */
     TIM_Channel trig_channel;       /**< Timer channel for trigger. */
-    volatile uint32_t distance_cm;  /**< Distance in centimeters. */
+    volatile float distance_cm;  /**< Distance in centimeters. */
 };
 
 /* Private functions ---------------------------------------------*/
@@ -48,7 +48,7 @@ void hc_sr04_init(struct us_sensor_str *us_sensor, TIM_HandleTypeDef *htim_echo,
   * @param  distance_us: Distance in microseconds.
   * @retval Distance in centimeters.
   */
-uint32_t hc_sr04_convert_us_to_cm(uint32_t distance_us);
+float hc_sr04_convert_us_to_cm(float distance_us);
 
 /**
   * @brief  Calculates the position based on two distance values.
@@ -56,7 +56,7 @@ uint32_t hc_sr04_convert_us_to_cm(uint32_t distance_us);
   * @param  dis2: Distance 2 value.
   * @retval Calculated position value.
   */
-int CalulatePosition(int dis1, int dis2);
+int CalulatePosition(float dis1, float dis2);
 
 #ifdef __cplusplus
 }
